requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
exceptiongroup==1.2.2
fastcore==1.7.0
fastlite==0.0.7
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.4.1
python-multipart==0.0.9
pythonanywhere==0.14.1
pythonanywhere-core==0.1.10
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
schema==0.7.7
shellingham==1.5.4
six==1.16.0
snakesay==0.10.2
sniffio==1.3.1
soupsieve==2.6
sqlite_minutils==3.36.0.post4
starlette==0.38.2
tabulate==0.9.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
